MDLSITE-4211 add travisfile and tests structure
There are actually quit a few things i'm not a fan of from my first
playing with bats - one was the fact there isn't a good way to get
output from failed tests! So because of that I have also included the
bats-assert and bats-error libraries.

See https://github.com/ztombol/bats-docs for details of the assertion
lib.